requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
websockets==12.0
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.0
httpx==0.28.1
pymupdf==1.24.0
chromadb==0.5.0
sentence-transformers==2.5.0
